move merge_timeseries function
