+++ cpp/examples/temporal_hybrid_dabm_osecir.cpp
@@ -33,7 +33,6 @@
 #include "memilio/geography/regions.h"
 #include "ode_secir/infection_state.h"
 #include "models/hybrid/conversion_functions.cpp"
-#include "models/hybrid/analyze_result.h"
 #include <cstddef>
 #include <vector>
 
@@ -214,6 +213,6 @@ int main()
     auto ts_ode = hybrid_sim.get_result_model2();
 
     //Print result time series
-    auto ts = mio::interpolate_simulation_result(mio::hybrid::merge_time_series(ts_abm, ts_ode));
+    auto ts = mio::interpolate_simulation_result(mio::merge_time_series(ts_abm, ts_ode));
     ts.print_table({"S", "E", "Ins", "Ins_confirmed", "Isy", "Isy_confirmed", "Isev", "Icri", "R", "D"});
 }
