+++ cpp/models/hybrid/CMakeLists.txt
@@ -3,7 +3,6 @@ add_library(hybrid
     temporal_hybrid_model.cpp
     conversion_functions.h
     conversion_functions.cpp
-    analyze_result.h
 )
 target_link_libraries(hybrid PUBLIC memilio)
 target_include_directories(hybrid PUBLIC
